requirements.txt
numpy
matplotlib>=2.2.4
scipy
jupyter
black[jupyter]==21.11b1
pytest
pytest-cov>=3.0.0
python-statemachine
